% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_management.R
\name{add_request}
\alias{add_request}
\title{Add request}
\usage{
add_request(expr, requests)
}
\arguments{
\item{expr}{An unevaluated expression}

\item{requests}{A vector of requests}
}
\description{
Add a request to the stack
}
