% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{agent_info}
\alias{agent_info}
\title{Agent info}
\usage{
agent_info(token, base_url)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}
}
\description{
Returns information about your agent
}
