% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{basic_loop}
\alias{basic_loop}
\title{Basic mining loop}
\usage{
basic_loop(token, base_url, ship_id, belt_waypoint = "X1-DF55-17335A")
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}

\item{belt_waypoint}{Waypoint id for resource waypoint}
}
\description{
Mine until cargo hold is full, then unload and repeat
}
