% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{buy_ship}
\alias{buy_ship}
\title{Buy a ship}
\usage{
buy_ship(token, base_url, ship_type, purchase_waypoint)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_type}{A ship type string}

\item{purchase_waypoint}{The waypoint to purchase from}
}
\description{
Buy a ship from the market
}
