% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{cmd}
\alias{cmd}
\title{Send command}
\usage{
cmd(endpoint = endpoint, method = method, requests = requests, ...)
}
\arguments{
\item{endpoint}{Endpoint of the API to call. Use "VAR" to represent variables}

\item{method}{Either "GET" or "POST"}

\item{requests}{A vector of pending calls}

\item{...}{Other arguments needed by child functions}
}
\description{
Generalized command for triggering actions
}
