% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{contracts}
\alias{contracts}
\title{Contract info}
\usage{
contracts(token, base_url)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}
}
\description{
Returns information about your active contracts
}
