% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{deliver}
\alias{deliver}
\title{Deliver contract items}
\usage{
deliver(token, base_url, ship_id, contract_id, contract_symbol, units)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}

\item{contract_id}{A contract id string}

\item{contract_symbol}{The market symbol for the item to sell}

\item{units}{Number of units to deliver. Integer.}
}
\description{
Deliver contract items to current location
}
