% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{dock}
\alias{dock}
\title{Dock a ship}
\usage{
dock(token, base_url, ship_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}
}
\description{
Docks a ship at its current location
}
