% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{extract}
\alias{extract}
\title{Extract resources}
\usage{
extract(ship_select, requests)
}
\arguments{
\item{requests}{A vector of pending requests}

\item{ship_symbol}{A ship symbol}
}
\description{
Extract resources
}
