% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{extraction_loop}
\alias{extraction_loop}
\title{Extraction loop}
\usage{
extraction_loop(token, base_url, ship_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{Ship ID string}

\item{resource_waypoint}{Waypoint ID for resource belt}
}
\description{
An action loop to extract, deliver, and sell resources
}
