% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{fly_and_dock}
\alias{fly_and_dock}
\title{Fly and dock}
\usage{
fly_and_dock(token, base_url, ship_id, waypoint_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}

\item{waypoint_id}{Destination waypoint id string}
}
\description{
Fly a ship to a waypoint and dock there
}
