% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_management.R
\name{game}
\alias{game}
\title{Start up}
\usage{
game(
  new_game = FALSE,
  token = NULL,
  symbol = NULL,
  faction = "COSMIC",
  pool = NULL
)
}
\arguments{
\item{new_game}{Force a new game?}

\item{token}{An optional agent token as provided by the \code{register} endpoint. If not provided the function will try to read from the local database. If neither are found, will create a new agent.}

\item{symbol}{Your agent name}

\item{faction}{Faction to join. See https://docs.spacetraders.io/game-concepts/agents-and-factions}
}
\description{
Initialize the basic environment needed for the game
}
