% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{market}
\alias{market}
\title{Get market info}
\usage{
market(token, base_url, system_id, waypoint_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{system_id}{A system id string}

\item{waypoint_id}{A waypoint id string}
}
\description{
Retrieve information about the market at a waypoint
}
