% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{navigate}
\alias{navigate}
\title{Extract until full}
\usage{
navigate(token, base_url, ship_id, waypoint_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}
}
\description{
Extract resources at the current location until the ship's cargo is full
}
