% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{new_agent}
\alias{new_agent}
\title{New agent}
\usage{
new_agent(base_url, symbol, faction = "COSMIC", overwrite = FALSE)
}
\arguments{
\item{base_url}{base url of the api}

\item{symbol}{Agent symbol/name}
}
\description{
Create a new agent
}
