% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{orbit}
\alias{orbit}
\title{Orbit}
\usage{
orbit(token, base_url, ship_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}
}
\description{
Orbit or undock a ship at its current location.
}
