% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{process_data}
\alias{process_data}
\title{Process data retrieved from API}
\usage{
process_data(token, base_url, d)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}
}
\description{
Process data retrieved from API
}
