% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_management.R
\name{process_stack}
\alias{process_stack}
\title{Process stack}
\usage{
process_stack(requests = NULL, token, base_url, pool)
}
\arguments{
\item{requests}{A vector of requests to process}
}
\description{
Execute the top request in the stack
}
