% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{refuel}
\alias{refuel}
\title{Refuel a ship}
\usage{
refuel(ship_select, requests)
}
\arguments{
\item{requests}{A vector of pending requests}

\item{ship_symbol}{A ship symbol}
}
\description{
Refuel a ship
}
