% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{run_swarm}
\alias{run_swarm}
\title{Run swarm}
\usage{
run_swarm(token, base_url)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}
}
\description{
Automate basic mining loop for all ships
}
