% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_management.R
\name{send_request}
\alias{send_request}
\title{Send request}
\usage{
send_request(method, token, base_url, endpoint, body = NULL)
}
\arguments{
\item{method}{One of "GET" or "POST"}

\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{endpoint}{API endpoint as a string}

\item{body}{Payload if needed, provided as a named list}
}
\description{
Sends a request to the spacetraders.io api
}
