% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{ship_cargo}
\alias{ship_cargo}
\title{Ship cargo}
\usage{
ship_cargo(token, base_url, ship_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}
}
\description{
Get information about the cargo of a ship
}
