% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_management.R
\name{tick}
\alias{tick}
\title{Tick}
\usage{
tick(
  game = NULL,
  stack = requests,
  input = NULL,
  output = NULL,
  check_time = NULL
)
}
\arguments{
\item{game}{An object produced by \code{game()}}

\item{check_time}{The time of the previous cycle}
}
\description{
A game tick
}
