% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\docType{data}
\name{token}
\alias{token}
\title{Constants}
\format{
An object of class \code{character} of length 1.
}
\usage{
token
}
\description{
token.txt should contain your spacetraders.io agent token
base_url is set to v2 of the spacetraders.io api
}
\keyword{datasets}
