% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{unload_cargo}
\alias{unload_cargo}
\title{Unload cargo}
\usage{
unload_cargo(token, base_url, ship_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}
}
\description{
Deliver contract items and sell non-contract items
}
