% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{update_data}
\alias{update_data}
\title{Query API}
\usage{
update_data(token, base_url, current_data)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}
}
\description{
Pull down data at once for processing
}
