% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{update_unless_error}
\alias{update_unless_error}
\title{Update unless error}
\usage{
update_unless_error(
  method,
  token,
  base_url,
  endpoint,
  d = NULL,
  json_body = NULL
)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}
}
\description{
Try to query the API. Return updated data if the query succeds
otherwise keep original data
}
