% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{deliver}
\alias{deliver}
\title{Deliver contract cargo}
\usage{
deliver(contract_select, requests)
}
\arguments{
\item{requests}{A vector of pending requests}

\item{ship_symbol}{A ship symbol}
}
\description{
Deliver contract cargo
}
