% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{dock}
\alias{dock}
\title{Dock a ship}
\usage{
dock(ship_symbol, requests, game)
}
\arguments{
\item{ship_symbol}{A ship symbol}

\item{requests}{A vector of pending requests}

\item{game}{A game object}
}
\description{
Dock a ship
}
