% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{extract}
\alias{extract}
\title{Extract resources}
\usage{
extract(token, base_url, ship_id)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}
}
\description{
Extract resources at the current location
}
