% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{orbit}
\alias{orbit}
\title{Orbit a ship}
\usage{
orbit(ship_symbol, requests, game)
}
\arguments{
\item{ship_symbol}{A ship symbol}

\item{requests}{A vector of pending requests}

\item{game}{A game object}
}
\description{
Orbit a ship
}
