% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{orbit}
\alias{orbit}
\title{Orbit a ship}
\usage{
orbit(ship_select, requests)
}
\arguments{
\item{requests}{A vector of pending requests}

\item{ship_symbol}{A ship symbol}
}
\description{
Orbit a ship
}
