% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sell}
\alias{sell}
\title{Sell an item}
\usage{
sell(token, base_url, ship_id, symbol, units)
}
\arguments{
\item{token}{API agent token}

\item{base_url}{base url of the api}

\item{ship_id}{A ship id string}

\item{symbol}{The market symbol for the item to sell}

\item{units}{Number of units to sell. Integer.}
}
\description{
Sell an item at the current location
}
